targets back to float()

+++ munglinker/model.py
@@ -611,7 +611,7 @@ class PyTorchNetwork(object):
             np_inputs, np_targets = next(val_generator)
 
             inputs = Variable(torch.from_numpy(np_inputs).float())
-            targets = Variable(torch.from_numpy(np_targets).long())
+            targets = Variable(torch.from_numpy(np_targets).float())
             if self.cuda:
                 inputs = inputs.cuda()
                 targets = targets.cuda()
@@ -811,7 +811,7 @@ class PyTorchNetwork(object):
             # return Torch Variables, which the model can however
             # specify in its prepare() function(s).
             inputs = Variable(torch.from_numpy(np_inputs).float())
-            targets = Variable(torch.from_numpy(np_targets).long())
+            targets = Variable(torch.from_numpy(np_targets).float())
             if self.cuda:
                 inputs = inputs.cuda()
                 targets = targets.cuda()
